// js/types.ts
// types.ts

export interface Tour {
    id: number;
    name: string;
    price: number;
    duration: string;
    rating: number;
    imageUrl: string;
}

export interface Hotel {
    id: number;
    name: string;
    location: string;
    price: number;
    rating: number;
    imageUrl: string;
}

export interface Blog {
    id: number;
    title: string;
    excerpt: string;
    author: string;
    date: string;
    imageUrl: string;
}

export interface FAQ {
    title: string;
    content: string;
}

export interface Slide {
    title: string;
    subtitle: string;
    bg: string;
}

export interface Tab {
    name: string;
}